Comment functions & set up example in scene tsx
