+++ src/Shapes/Box.tsx
@@ -1,21 +1,18 @@
 import React from 'react';
-import { SphereProps, useSphere } from '@react-three/cannon';
+import { BoxProps, useBox } from '@react-three/cannon';
 
-interface CustomSphere extends SphereProps{
-  color: string
+interface CustomBox extends BoxProps {
+  color: string;
 }
 /**
- * CollisionObject component renders an object with collision detection.
+ * Box component renders a box with collision detection.
  * This component accepts props that determine its size, position, type, and mass,
  * and a callback to be called upon a collision.
- *
- * TODO: Support passing props for multiple shapes rather than just a sphere...
  * @param props args: Array<number>, mass: number, position: Triplet, type: string, onCollide: Function
  * @returns
  */
-export default function Sphere(props: CustomSphere) {
-
-  const [collisionRef] = useSphere(() => ({
+export default function Box(props: CustomBox) {
+  const [collisionRef] = useBox(() => ({
     args: props.args,
     mass: props.mass,
     position: props.position,
@@ -26,7 +23,7 @@ export default function Sphere(props: CustomSphere) {
 
   return (
     <mesh ref={collisionRef}>
-      <sphereBufferGeometry args={props.args} />
+      <boxBufferGeometry args={props.args} />
       <meshPhongMaterial color={props.color} />
     </mesh>
   );
