+++ src/Shapes/Cylinder.tsx
@@ -5,11 +5,9 @@ interface CustomCylinder extends CylinderProps {
   color: string;
 }
 /**
- * CollisionObject component renders an object with collision detection.
+ * Cylinder component renders a cylinder with collision detection.
  * This component accepts props that determine its size, position, type, and mass,
  * and a callback to be called upon a collision.
- *
- * TODO: Support passing props for multiple shapes rather than just a sphere...
  * @param props args: Array<number>, mass: number, position: Triplet, type: string, onCollide: Function
  * @returns
  */
